Move goal-general function from CharacterInverntoryActions to GoalActions.
